disable clearReportNotifications on android
