--- src/libs/Notification/clearReportNotifications/index.native.ts
@@ -1,5 +1,0 @@
-import PushNotification from '@libs/Notification/PushNotification';
-import ClearReportNotifications from './types';
-
-const clearReportNotifications: ClearReportNotifications = PushNotification.clearReportNotifications;
-export default clearReportNotifications;
